requirements.txt
black
build
coverage
flake8
isort
mypy
PyGithub>=2,<3
pytest
PyYAML>=6,<7
ruff
setuptools
twine
types-PyYAML
yamllint
